pyproject.toml
[project]
name = "fineasy-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "brapi>=1.2.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
]
